remove SetNotifiesStage into its own file
